Moved the JSBSim module to a separate file
